Finished the tutorial
The app works but some things aren't working correctly
Need to fix the description not showing up when updating.
Need to see why when you update, the new information isn't shown correctly.
Need to fix the bootstrap so that you can still nav to other page when browser is resized.